functions/extract-arxiv/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
arxiv==2.*
